Support ssl connect without specifying options
Previously you had to pass a non-empty list in {ssl, list()} in order to get
brod to connect to Kafka using ssl. This is not needed, the presence of
{ssl, list()} in the client options should be enough to deduce that we should
connect to Kafka using ssl. Having this requirement makes it awkward to
configure a brod client on a host using the CA certificates provided by the host
environment and connects to an endpoint with a certificate that can be verified
as such.